[breaking] align api
